pyproject.toml
[project]
name = "py-websites-scraper"
version = "0.1.0"
description = "Mutliple we scraper using aiohttp to make it easy to scrape multiple URLs"
readme = "README.md"
requires-python = ">=3.8"
authors = [
  { name = "Hilman Ramadhan", email = "hilmanhgb@gmail.com" }
]
license = { text = "MIT" }
dependencies = [
  "aiohttp>=3.12.2",
  "readability-lxml>=0.8.4"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "black",
  "isort",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.urls]
Homepage = "https://github.com/hilmanski/py-websites-scraper"
Issues = "https://github.com/hilmanski/py-websites-scraper"